requirements-api.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
torch>=2.1.0
transformers>=4.40.0
Pillow>=10.0.0
numpy>=1.25.0